fix(sidecar,cli): move the dispatcher into the CLI and build Windows on macOS
Each per-platform sidecar sub-package previously built from the parent's source through a cross-package filesystem path. The argv dispatcher now lives in vitest-agent-cli as a clean export, and each sub-package has its own entry that imports it — an ordinary package import, no file-boundary reach.

The parent package drops its dispatcher source; it keeps the launcher shim, the rslib build, and the optional dependencies on the five children. The dispatcher test moves to the CLI package alongside the code it now covers.

The cross-platform build job moves to a macOS runner: building the Windows single-executable extracts a zip archive, which the macOS tar reads but the Linux tar cannot. The five child builds also run serially to avoid a race on the shared runtime-download cache.
