+++ packages/cli/__test__/sidecar-dispatch.test.ts
@@ -2,12 +2,12 @@ import { mkdtempSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { dispatch } from "../src/bin.js";
+import { dispatch } from "../src/lib/sidecar-dispatch.js";
 
 let cwd: string;
 
 beforeEach(() => {
-	cwd = mkdtempSync(join(tmpdir(), "sidecar-bin-"));
+	cwd = mkdtempSync(join(tmpdir(), "sidecar-dispatch-"));
 });
 
 afterEach(() => {
